add profiling test
